refactor: move native window gin converter (#19577)
* move native window gin converter

* extract to new file
