+++ filenames.gni
@@ -471,6 +471,7 @@ filenames = {
     "shell/common/gin_converters/image_converter_gin_adapter.h",
     "shell/common/gin_converters/message_box_converter.cc",
     "shell/common/gin_converters/message_box_converter.h",
+    "shell/common/gin_converters/native_window_converter.h",
     "shell/common/gin_converters/net_converter_gin_adapter.h",
     "shell/common/gin_converters/string16_converter.h",
     "shell/common/gin_util.h",
@@ -502,6 +503,7 @@ filenames = {
     "shell/common/native_mate_converters/image_converter.cc",
     "shell/common/native_mate_converters/image_converter.h",
     "shell/common/native_mate_converters/map_converter.h",
+    "shell/common/native_mate_converters/native_window_converter.h",
     "shell/common/native_mate_converters/net_converter.cc",
     "shell/common/native_mate_converters/net_converter.h",
     "shell/common/native_mate_converters/network_converter.cc",
