+++ shell/browser/api/atom_api_screen.cc
@@ -10,9 +10,9 @@
 #include "base/bind.h"
 #include "native_mate/dictionary.h"
 #include "native_mate/object_template_builder.h"
-#include "shell/browser/api/atom_api_browser_window.h"
 #include "shell/browser/browser.h"
 #include "shell/common/native_mate_converters/gfx_converter.h"
+#include "shell/common/native_mate_converters/native_window_converter.h"
 #include "shell/common/node_includes.h"
 #include "ui/display/display.h"
 #include "ui/display/screen.h"
