+++ shell/common/gin_converters/message_box_converter.h
@@ -6,20 +6,9 @@
 #define SHELL_COMMON_GIN_CONVERTERS_MESSAGE_BOX_CONVERTER_H_
 
 #include "gin/converter.h"
-#include "shell/browser/api/atom_api_browser_window.h"
 #include "shell/browser/ui/message_box.h"
 
 namespace gin {
-
-// TODO(deermichel): remove adapter after removing mate
-template <>
-struct Converter<electron::NativeWindow*> {
-  static bool FromV8(v8::Isolate* isolate,
-                     v8::Local<v8::Value> val,
-                     electron::NativeWindow** out) {
-    return mate::ConvertFromV8(isolate, val, out);
-  }
-};
 
 template <>
 struct Converter<electron::MessageBoxSettings> {
